+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-06-28 13:59:36
-// Commit: chore(theme): update dependencies and cleanup unused imports in theme
+// Updated at: 2026-06-28 15:46:33
+// Commit: feat(theme): add async theme stream provider
 
 class themeComponentState {
-  final String commitRef = '2392';
-  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
-  final DateTime timestamp = DateTime.parse('2026-06-28T13:59:36+05:00');
+  final String commitRef = '2394';
+  final String lastMessage = 'feat(theme): add async theme stream provider';
+  final DateTime timestamp = DateTime.parse('2026-06-28T15:46:33+05:00');
 }